+++ src/main/java/team/hackerping/nanuri/article/presentation/dto/ArticleRequest.java
@@ -3,7 +3,7 @@ package team.hackerping.nanuri.article.presentation.dto;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
-import team.hackerping.nanuri.article.application.RegisterArticleCommand;
+import team.hackerping.nanuri.article.application.command.RegisterArticleCommand;
 import team.hackerping.nanuri.article.domain.ArticleStatus;
 import team.hackerping.nanuri.article.domain.ItemCategory;
 import team.hackerping.nanuri.article.domain.ShareType;
